refactor: move log request merging to filter utility
